requirements.txt
python-dotenv==0.20.0
Telethon==1.24.0
APScheduler==3.9.1
